+++ fabric/src/main/java/io/github/sakurawald/module/initializer/color/sign/structure/SignCache.java
@@ -1,10 +1,13 @@
 package io.github.sakurawald.module.initializer.color.sign.structure;
 
 import lombok.Data;
+import lombok.With;
 
 import java.util.List;
 
 @Data
+@With
 public class SignCache {
-    final List<String> lines;
+    final List<String> frontLines;
+    final List<String> backLines;
 }