+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - feature: new supported `MC 1.20` version!
 - [chat.history] fix: make it working in `offline-mode` server.
+- [color.sign] fix: should save both `front texts` and `back texts`.
 - [command_permission] feature: now we can detect `in-complete command` or `in-correct argument` when using `/command-permission describe <cmd>` command.
 - [system_message] feature: add debug logging when replacing a translatable text.